fix: 修复系统用户详情中测试连接性
