+++ src/components/AccountListTable/index.vue
@@ -8,10 +8,10 @@
 
 <script>
 import ListTable from '@/components/ListTable/index'
-import { ActionsFormatter, ChoicesFormatter, DetailFormatter } from '@/components/TableFormatters'
+import { ActionsFormatter, DetailFormatter } from '@/components/TableFormatters'
 import ShowSecretInfo from './ShowSecretInfo'
 import UpdateSecretInfo from './UpdateSecretInfo'
-import { toSafeLocalDateStr } from '@/utils/common'
+import { connectivityMeta } from './const'
 
 export default {
   name: 'Detail',
@@ -77,33 +77,7 @@ export default {
           version: {
             width: '70px'
           },
-          connectivity: {
-            label: this.$t('assets.Reachable'),
-            formatter: ChoicesFormatter,
-            formatterArgs: {
-              iconChoices: {
-                ok: 'fa-check text-primary',
-                failed: 'fa-times text-danger',
-                unknown: 'fa-circle text-warning'
-              },
-              hasTips: true,
-              getTips: ({ row, cellValue }) => {
-                const mapper = {
-                  'ok': this.$t('assets.Reachable'),
-                  'failed': this.$t('assets.Unreachable'),
-                  'unknown': this.$t('assets.Unknown')
-                }
-                let tips = mapper[cellValue]
-                if (row['date_verified']) {
-                  const datetime = toSafeLocalDateStr(row['date_verified'])
-                  tips += '<br> ' + datetime
-                }
-                return tips
-              }
-            },
-            width: '90px',
-            align: 'center'
-          },
+          connectivity: connectivityMeta,
           actions: {
             formatter: ActionsFormatter,
             formatterArgs: {
